removing unnecessary files
