+++ src/App.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom'
 import './App.css';
 import SignUpForm from './components/SignUpForm';
-import ServiceRequestForm from './components/ServiceRequestForm1';
-import AdditionalInformationForm from './components/AdditionalInformation1';
+import ServiceRequestForm from './components/ServiceRequestForm';
+import AdditionalInformationForm from './components/AdditionalInformation';
 import SignInForm from './components/SignInForm';
 import ResetPassword from './components/ResetPassword';
 import withFormik from './components/withFormik';
