+++ src/components/ServiceRequestForm.jsx
@@ -1,92 +1,205 @@
 import React, { useState, useEffect } from "react";
-import { Formik, Form, Field, connect } from "formik";
+import { Form, Field, connect } from "formik";
 import axios from "axios"
-import * as Yup from "yup";
 import ErrorMsg from "./ErrorMessage";
 import FormContainer from './FormContainer';
 import RequestTypeField from "./RequestTypeField";
+import RequestSubTypeField from "./RequestSubTypeField";
 import RequestPetTypeField from "./RequestPetTypeField";
+import QueryString from 'query-string';
+
+import _ from 'lodash'
+import GenericTypeField from "./genericTypeField";
 //import PetTypes from "./pettypes.json";
 //import AnimalBreeds from "./animalbreeds.json";
 //import AnimalColors from "./animalcolors.json"
+import Model from './Model'
 
+const {categoryId} = QueryString.parse(window.location.search);
 
-const getSubCategories = (categories, categoryId) => {
+const getSubCategories1 = (categories, categoryId) => {
 	var category = categories.find(category => category.id === categoryId);
 	return category ? category.types : [];
 };
-const getAnimalSubCategories = (AnimalBreeds, animalId) => {
+const getSubCategories = (categories, categoryName) => {
+	var category = categories.find(category => category.name.toLowerCase() === categoryName);
+	return category ? category.types : [];
+};
+const getNote = (subCategories, name) => {
+	var type = subCategories.find(subcategoryname => subcategoryname.name.toLowerCase() === name);
+	return type ? type.note : [];
+};
+const getAnimalSubCategories = (AnimalBreeds, animalName) => {
 	//console.log('--inside getAnimalSubCategories--');
 
-	var animalCats = AnimalBreeds.find(animal => animal.id === animalId);
+	var animalCats = AnimalBreeds.find(animal => animal.animal.toLowerCase() === animalName);
 	return animalCats ? animalCats : [];
 
 };
+const getID = (categories, categoryName) => {
 
 
-const ServiceRequestForm = props => {
+	var category = categories.find(category => category.name.toLowerCase() === categoryName);
+	return category ? category.id : [];
+};
+
+// const getSubTypeID = (categories, categoryName) => {
+// 	var category = categories.find(category => category.name=== categoryName);
+// 	return category ? category.id : [];
+// };
+
+
+const ServiceRequestForm = (props, errors, touched) => {
+
+	const requestType_petAndAnimalIssue = 'Pets and Animals Issue';
+	const petAndAnimalIssueID_OtherAnimalComplaint = 'Other animal complaint';
+
+	const requestType_WebSiteIssue = 'Website Issue';
+	const subCategory_OtherWebsiteProblem = 'Other website problem';
+
+	const requestType_TrashRecycleIssue = 'Trash and Recycling Issue';
+
+
+	const requestType_WaterandSewerIssues = 'Water and Sewer Issues';
+	const subCategory_CanOrLidLostDamaged = 'Can or lid lost or damaged';
+	const subCategory_PropertyDamangeDuringCollecttion = 'Property damage during collection';
+	const subCategory_RecyclingNotCollected = 'Recycling not collected';
+	const subCategory_RequestToStartNewCollection = 'Request to start new collection';
+	const subCategory_TrashNotCollected = 'Trash not collected';
+	const subCategory_YardWasteNotCollected = 'Yard waste not collected';
+
+
+
+
+	const requestType_RoadSidewalkIssue = 'Roads and Sidewalks Issue';
+	const subCategory_IcyConditions = 'Icy conditions';
+
+	const subCategory_SewerIssues = 'Sewer issues';
+	const subCategory_StormWaterIssues = 'Stormwater issues';
+	const subCategory_WaterSupplyIssues = 'Water supply issues';
+
+
+	const petTypeCat = 'cat';
+	const petTypeDog = 'dog';
+	const petTypeHorse = 'horse';
+	const petType_Others = 'other';
 	const [Categories, setData] = useState([]);
 	const [PetTypes, setPetTypes] = useState([]);
 	const [AnimalBreeds, setAnimalBreeds] = useState([]);
 	const [AnimalColors, setAnimalColors] = useState([]);
 	const [OtherAnimalTypes, setOtherAnimalTypes] = useState([]);
 	const [subCategories, setSubCategories] = useState([]);
-	//const [breeds] = useState(AnimalBreeds);
+	const [notes, setNotes] = useState();
 	const [animalSubCategories, setAnimalSubCategories] = useState([]);
 	const [animalSex, setAnimalSex] = useState([]);
 
+	try {
+		useEffect(() => {
+			console.log('test--');
 
-	useEffect(() => {
-		const fetchData = async () => {
-			const result = await axios(
-				'//dev.baltimorecountymd.gov/sebin/q/m/categories.json',
-			);
-			const resultPetTypes = await axios(
-				'//dev.baltimorecountymd.gov/sebin/m/a/pet-types.json',
-			);
-			const resultAnimalBreeds = await axios(
-				'//dev.baltimorecountymd.gov/sebin/y/a/animal-breeds.json',
-			);
-			const resultAnimalColors = await axios(
-				'//dev.baltimorecountymd.gov/sebin/u/u/animal-colors.json',
-			);
-			const resultAnimalTypes = await axios(
-				'//dev.baltimorecountymd.gov/sebin/a/e/animal-types.json',
-			);
-			setData(result.data);
-			setPetTypes(resultPetTypes.data);
-			setAnimalBreeds(resultAnimalBreeds.data);
-			setAnimalColors(resultAnimalColors.data);
-			setOtherAnimalTypes(resultAnimalTypes.data);
-		};
 
-		fetchData();
-	}, []);
+			const fetchData = async () => {
+				const result = await axios(
+					'//dev.baltimorecountymd.gov/sebin/q/m/categories.json',
+				);
+				const resultPetTypes = await axios(
+					'//dev.baltimorecountymd.gov/sebin/m/a/pet-types.json',
+				);
+				const resultAnimalBreeds = await axios(
+					'//dev.baltimorecountymd.gov/sebin/y/a/animal-breeds.json',
+				);
+				const resultAnimalColors = await axios(
+					'//dev.baltimorecountymd.gov/sebin/u/u/animal-colors.json',
+				);
+				const resultAnimalTypes = await axios(
+					'//dev.baltimorecountymd.gov/sebin/a/e/animal-types.json',
+				);
+				setData(result.data);
+				setPetTypes(resultPetTypes.data);
+				setAnimalBreeds(resultAnimalBreeds.data);
+				setAnimalColors(resultAnimalColors.data);
+				setOtherAnimalTypes(resultAnimalTypes.data);
+
+			};
+
+			fetchData();
+		}, []);
+	}
+	catch (ex) {
+
+		console.log(ex);
+	}
 
 
 	const handleServiceRequestChange = (changeEvent) => {
-		console.log('++++++++++++++handleServiceRequestChange+++++++++++++');
-		const { value } = changeEvent.currentTarget;
-		console.log('RequestChange value:' + value);
-		const subCategories = getSubCategories(Categories, parseInt(value));
+
+		console.log('-------------handleServiceRequestChange-------------');
+		const value = changeEvent.currentTarget.value.toLowerCase();
+		let ID = getID(Categories, value)
+
+		props.formik.setFieldValue('requestTypeID', ID);
+		const subCategories = getSubCategories(Categories, value ? value : value);
+		//const subCategories = getSubCategories(Categories, parseInt(categoryID) ? parseInt(categoryID) : categoryID);
 		setSubCategories(subCategories);
 	};
-	const handleServicePetChange = (changeEvent) => {
-		console.log('===============handleServicePetChange===============');
-		const { value } = changeEvent.currentTarget;
-		console.log('pet value:' + value);
-		const subBreeds = getAnimalSubCategories(AnimalBreeds, parseInt(value));
-		setAnimalSubCategories(subBreeds.breeds);
-		console.log(subBreeds.sex);
-		setAnimalSex(subBreeds.sex)
+	const handleServiceSubRequestChange = (changeEvent) => {
+
+		const value = changeEvent.currentTarget.value.toLowerCase();
+		let ID = getID(subCategories, value)
+		const notes = getNote(subCategories, value);
+		setNotes(<div><p>{notes}</p></div>);
+
+		props.formik.setFieldValue('subRequestTypeID', ID);
 
 	};
+	const handleServicePetChange = (changeEvent) => {
 
-	const getAnimalSex = (animalId) => {
-		console.log('getAnimalSex');
-		var animalCats = AnimalBreeds.find(animal => animal.id === parseInt(animalId));
+		let value = changeEvent.currentTarget.value.toLowerCase();
+		let ID = getID(PetTypes, value)
+		const subBreeds = getAnimalSubCategories(AnimalBreeds, value);
+		setAnimalSubCategories(subBreeds.breeds);
+		setAnimalSex(subBreeds.sex);
+		props.formik.setFieldValue('petTypeID', ID);
+
+
+	};
+	const handleAnimalColorChange = (changeEvent) => {
+
+		let value = changeEvent.currentTarget.value.toLowerCase();
+		let ID = getID(AnimalColors, value);
+		props.formik.setFieldValue('animalColorTypeID', ID);
+		//props.formik.setFieldValue('animalColorType', value);
+	}
+
+	const handleOtherPetTypeChange = (changeEvent) => {
+
+		let value = changeEvent.currentTarget.value.toLowerCase();
+		let ID = getID(OtherAnimalTypes, value);
+		props.formik.setFieldValue('otherAnimalTypesID', ID);
+		//props.formik.setFieldValue('otherAnimalTypes', value);
+	}
+	const handlePetSexChange = (changeEvent) => {
+		let value = changeEvent.currentTarget.value.toLowerCase();
+		let ID = getID(animalSex, value);
+		props.formik.setFieldValue('sexTypeID', ID);
+		//props.formik.setFieldValue('sexType', value);
+	}
+
+
+
+	const handleAnimalBreedChange = (changeEvent) => {
+
+		let value = changeEvent.currentTarget.value.toLowerCase();
+		let ID = getID(animalSubCategories, value);
+
+		props.formik.setFieldValue('animalBreedID', ID);
+		//props.formik.setFieldValue('animalBreed', value);
+	}
+
+	const checkPetType = (value) => {
+		value = value.toLowerCase();
+		var animalCats = AnimalBreeds.find(animal => animal.animal === value);
 		if (animalCats !== undefined) {
-			setAnimalSex(animalCats.sex);
 			return true;
 		}
 		else {
@@ -95,279 +208,399 @@ const ServiceRequestForm = props => {
 
 
 	}
-	const callRegisterForm = () => {
-		console.log('---callregisterForm---');
-		props.history.push("/SignUpForm");
-	}
 
+
+	const buttonShowHideValidation = () => {
+
+		console.log('iniside buttonShowHideValidation');
+		//console.log(props.formik.values.Id);
+		if (props.formik.values.requestType !== '' && props.formik.values.subRequestType !== '') {
+			if (props.formik.values.requestType.toLowerCase() === requestType_RoadSidewalkIssue.toLowerCase()
+				&& props.formik.values.subRequestType.toLowerCase() === subCategory_IcyConditions.toLowerCase()) {
+				return true;
+			}
+
+			else if ((rest.formik.values['requestType'].toLowerCase() === requestType_WaterandSewerIssues.toLowerCase())
+				&& (rest.formik.values['subRequestType'].toLowerCase() === subCategory_SewerIssues.toLowerCase() ||
+					rest.formik.values['subRequestType'].toLowerCase() === subCategory_StormWaterIssues.toLowerCase() ||
+					rest.formik.values['subRequestType'].toLowerCase() === subCategory_WaterSupplyIssues.toLowerCase()
+				)) {
+				return true;
+			}
+			else if ((rest.formik.values['requestType'].toLowerCase() === requestType_TrashRecycleIssue.toLowerCase())
+				&& (rest.formik.values['subRequestType'].toLowerCase() === subCategory_CanOrLidLostDamaged.toLowerCase() ||
+					rest.formik.values['subRequestType'].toLowerCase() === subCategory_PropertyDamangeDuringCollecttion.toLowerCase() ||
+					rest.formik.values['subRequestType'].toLowerCase() === subCategory_RecyclingNotCollected.toLowerCase() ||
+					rest.formik.values['subRequestType'].toLowerCase() === subCategory_RequestToStartNewCollection.toLowerCase() ||
+					rest.formik.values['subRequestType'].toLowerCase() === subCategory_TrashNotCollected.toLowerCase() ||
+					rest.formik.values['subRequestType'].toLowerCase() === subCategory_YardWasteNotCollected.toLowerCase()
+				)) {
+
+				return true;
+			}
+			else if ((rest.formik.values['requestType'].toLowerCase() === requestType_WebSiteIssue.toLowerCase()
+				&& rest.formik.values['subRequestType'].toLowerCase() === subCategory_OtherWebsiteProblem.toLowerCase())) {
+				return true;
+			}
+
+			else if (rest.formik.values['requestType'].toLowerCase() === requestType_petAndAnimalIssue.toLowerCase()
+				&& rest.formik.values['subRequestType'] !== ''
+				&& rest.formik.values['petType'] !== ''
+				&& (rest.formik.values['petType'].toLowerCase() !== petTypeHorse.toLowerCase()
+					&& rest.formik.values['petType'].toLowerCase() !== petType_Others.toLowerCase())
+
+				&& rest.formik.values['animalColorType'] === '') {
+
+				return true;
+			}
+			else if (rest.formik.values['requestType'].toLowerCase() === requestType_petAndAnimalIssue.toLowerCase()
+				&& rest.formik.values['subRequestType'] !== ''
+				&& (rest.formik.values['petType'] !== ''
+					&& (rest.formik.values['petType'].toLowerCase() === petType_Others.toLowerCase()
+						&& rest.formik.values['otherAnimalTypes'] === ''))) {
+
+				return true;
+			}
+			else {
+
+				return false;
+			}
+
+
+		}
+		else {
+
+			return true;
+		}
+
+
+	}
 	const callSignInForm = () => {
-		console.log('---callregisterForm---');
+
 		props.history.push("/SignInForm");
 	}
+	const callRegisterForm = () => {
 
+		props.history.push("/SignUpForm");
+	}
+	const goToNextPage = () => {
+		props.history.push('/ProviderDetails');
+	}
+	const { values, isSubmitting, ...rest } = props;
+
+	const routURLID = () => {
+		var urlParts =  window.location.href.split('categoryId=');
+		return urlParts[1];
+	}
+
+	//console.log('************************************************');
+	//console.log('requestType:' + rest.formik.values.requestType);
+	console.log('sexType:' + rest.formik.values.sexType);
+	//console.log('************************************************');
+	let displayButton = buttonShowHideValidation(props);
 	return (
 
 
 		<FormContainer title="How Can We Help?">
-			<Formik
-				initialValues={{
-					requestType: '',
-					subRequestType: '',
-					petType: '',
-					animalBreed: '',
-					sexType: '',
-					animalColorType: '',
-					otherAnimalTypes: ''
-				}}
-				validationSchema={Yup.object().shape({
-					requestType: Yup.string().required('Request Category is required'),
-					subRequestType: Yup.string().required('Sub Category is required'),
-					petType: Yup.string().required('Pet Type is required'),
-					animalColorType: Yup.string().required('Primary Animal Color is required'),
-				})}
 
-				onSubmit={(values, { setSubmitting }) => {
-					alert(JSON.stringify(values, null, 2));
-					//handleSubmitTest();
-					setSubmitting(false);
-				}}
-			>
+			<Form>
+				<label htmlFor="requestType"
+					className={
+						rest.formik.errors.requestType && rest.formik.touched.requestType ? "input-feedback" : "text-label"}
+				>Request Category</label>
+
+				<RequestTypeField
+					component="select"
+					name="requestType"
+					formikProps={rest}
+					onChange={handleServiceRequestChange}
+					onLoad={routURLID}
+					//value ={categoryId}
+					value={props.formik.values.name}
+				>
+					<option key='default' value=''>--Please select a category--</option>
+					{Categories.map(category => (
+						<option key={category.id} value={category.name}>{category.name}</option>
+					))}
+				</RequestTypeField>
+				<div className="input-feedback">
+					<ErrorMsg
+						errormessage={rest.formik.errors.requestType}
+						touched={rest.formik.touched.requestType} />
+				</div>
 				{
-					(props) => {
-						const { values, isSubmitting, errors, touched, ...rest } = props;
-						console.log('requestType:' + values['requestType']); //1010188
-						//	console.log('subRequestType:' + values['subRequestType']); //1010188
-						//	console.log('petType:' + values['petType']); //1010188
-						return (
+					rest.formik.values['requestType'] !== '' ?
+						<div>
+							<label name="subRequestType" htmlFor="subRequestType"
+								className={
+									rest.formik.errors.subRequestType && rest.formik.touched.subRequestType ? "input-feedback" : "text-label"}
+							>
+								Request Sub-Category
+							</label>
+							<RequestSubTypeField
+								component="select"
+								name="subRequestType"
+								formikProps={rest}
+								onChange={handleServiceSubRequestChange}
+								value={props.formik.values.name}
+							>
+								<option key='default' value=''>--Please select a sub-category--</option>
+								{subCategories.map(category => (
+									<option key={category.id} value={category.name}>{category.name}</option>
+								))}
 
-							<Form >
-								<label htmlFor="requestType"
-									className={
-										errors.requestType && touched.requestType ? "input-feedback" : "text-label"}
-								>Request Category</label>
-								<RequestTypeField
-									component="select"
-									name="requestType"
-									formikProps={rest}
-									onChange={handleServiceRequestChange}
-									className={`text-select ${errors.requestType && touched.requestType ? "error" : ""}`}
-								>
-									<option key='default' value=''>--Please select a category--</option>
-									{Categories.map(category => (
-										<option key={category.id} value={category.id}>{category.name}</option>
-									))}
-								</RequestTypeField>
-								<div className="input-feedback">
-									<ErrorMsg
-										errormessage={errors.requestType}
-										touched={touched.requestType} />
-								</div>
-								{
-									values['requestType'] !== '' ?
-										<div>
-											<label name="subRequestType" htmlFor="subRequestType"
-												className={
-													errors.subRequestType && touched.subRequestType ? "input-feedback" : "text-label"}
-											>
-												Request Sub-Category
-											</label>
-											<Field component="select"
-												name="subRequestType"
-												className={`text-select ${errors.subRequestType && touched.subRequestType ? "error" : ""}`}
-											>
-												<option key='default' value=''>--Please Select a sub-category--</option>;
-												{subCategories.map(category => (
-													<option key={category.id} value={category.id}>{category.name}</option>
-												))}
-											</Field>
-											<div className="input-feedback">
-												<ErrorMsg
-													errormessage={errors.subRequestType}
-													touched={touched.subRequestType} />
-											</div>
-										</div>
-										: null
-								}
-
-								{ /* Can or lid lost or damaged*/
-									values['subRequestType'] === '1011083' ?
-										<div>
-											The County will only replace damaged cans or lids with evidence of hauler negligence. After submitting your report, please email any evidence to solidwaste@baltimorecountymd.gov and reference your request ID number.
-										</div>
-										: null
-								}
-								{ /* icy condition*/
-									values['subRequestType'] === '1011069' ?
-										<div>
-											Due to the safety risk posed by ice in the roadway, we cannot take your report online.
-											Please call the Department of Public Works immediately at 410-887-0000 to ensure we obtain the necessary information to address the issue as soon as possible.
-										</div>
-										: null
-								}
-								{
-									values['subRequestType'] === '1010186' ?
-										<div>
-
-											The Police Department is responsible for investigating all animal cruelty incidents. If you witness a person inflicting harm on an animal, call 911 immediately.
-											To report animal cruelty when immediate police intervention is not required, call the police non-emergency line at 410-887-2222. Please don't use this form to report animal cruelty.
-										</div>
-										: null
-								}
-								{
-									values['requestType'] === '1010169' && values['subRequestType'] !== '' ?
-										<div>
-											<label htmlFor="petType" name="petType"
-												className={
-													errors.petType && touched.petType ? "input-feedback" : "text-label"}
-											>Pet Type</label>
-											<RequestPetTypeField
-												component="select"
-												name="petType"
-												formikProps={rest}
-												onChange={handleServicePetChange}
-												className={errors.petType && touched.petType ? "text-select error" : null}
-											>
-												<option key='default' value=''>--Please select a pet type--</option>
-												{PetTypes.map(petType => (
-													<option key={petType.id} value={petType.id}>{petType.name}</option>
-												))}
-											</RequestPetTypeField>
-											<div className="input-feedback">
-												{<ErrorMsg
-													errormessage={errors.petType}
-													touched={touched.petType} />}
-											</div>
-										</div>
-										: null
-								}
-								{
-
-									(values['petType'] === '1010191' && values['subRequestType'] !== '') ?
-										<div>
-											<label htmlFor="otherAnimalTypes"
-												className={
-													errors.otherAnimalTypes && touched.otherAnimalTypes ? "input-feedback" : "text-label"}
-											>Other pet type
-											</label>
-											<Field
-												component="select"
-												name="otherAnimalTypes"
-												className={errors.otherAnimalTypes && touched.otherAnimalTypes ? "text-select error" : null}
-											>
-												<option key='default' value=''>--Please select an "other" pet type--</option>
-												{OtherAnimalTypes.map(OtherAnimalType => (
-													<option key={OtherAnimalType.id} value={OtherAnimalType.id}>{OtherAnimalType.name}</option>
-												))}
-
-											</Field>
-
-											<div className="input-feedback">
-												{<ErrorMsg
-													errormessage={errors.otherAnimalTypes}
-													touched={touched.otherAnimalTypes} />}
-											</div>
-										</div>
-										: null
-								}
-								{
-									((values['requestType'] === '1010169') && values['subRequestType'] === '1010175') && getAnimalSex(values['petType']) ?
-										<div>
-											<label htmlFor="sexType"
-												className={
-													errors.sexType && touched.sexType ? "input-feedback" : "text-label"}
-											>Pet Sex (optional)
-											</label>
-											<Field
-												component="select"
-												name="sexType"
-												className={errors.sexType && touched.sexType ? "text-select error" : null}
-											>
-												<option key='default' value=''>--Please select a pet sex--</option>
-												{animalSex.map(petSex => (
-													<option key={petSex.id} value={petSex.id}>{petSex.name}</option>
-												))}
-
-											</Field>
-
-											<div className="input-feedback">
-												{<ErrorMsg
-													errormessage={errors.sexType}
-													touched={touched.sexType} />}
-											</div>
-										</div>
-										: null
-								}
-
-								{
-
-									(values['requestType'] === '1010169' && values['subRequestType'] === '1010175')
-										&& (values['petType'] === '1010188' || values['petType'] === '1010189') ?
-										<div>
-											<label htmlFor="animalColorType"
-												className={
-													errors.animalColorType && touched.animalColorType ? "input-feedback" : "text-label"}
-											>Primary Animal Color
-											</label>
-
-											<Field
-												component="select"
-												name="animalColorType"
-												className={errors.animalColorType && touched.animalColorType ? "text-select error" : null}
-											>
-												<option key='default' value=''>--Please select the primary color of the animal--</option>
-
-												{AnimalColors.map(animalColorType => (
-													<option key={animalColorType.id} value={animalColorType.id}>{animalColorType.name}</option>
-												))}
-
-											</Field>
-											<div className="input-feedback">
-												{<ErrorMsg
-													errormessage={errors.animalColorType}
-													touched={touched.animalColorType} />}
-											</div>
-										</div>
-										: null
-
-								}
-
-								<br />
-								{
-									((values['requestType'] === '1010169' && values['subRequestType'] === '1010175') &&
-										animalSubCategories !== undefined) && (values['petType'] === '1010188' || values['petType'] === '1010189') ?
-										<div>
-											<label htmlFor="animalBreed"
-												className={
-													errors.animalBreed && touched.animalBreed ? "input-feedback" : "text-label"}
-											>Primary Animal Breed(optional)
-											</label>
-											<Field
-												component="select"
-												name="animalBreed"
-												className={errors.animalBreed && touched.animalBreed ? "text-select error" : null}
-											>
-												<option key='default' value=''>--Please select the primary breed of the animal--</option>
-												{animalSubCategories.map(animalBreed => (
-													<option key={animalBreed.id} value={animalBreed.id}>{animalBreed.name}</option>
-												))}
-											</Field>
-											<div className="input-feedback">
-												{<ErrorMsg
-													errormessage={errors.animalBreed}
-													touched={touched.animalBreed} />}
-
-											</div>
-										</div>
-										: null
-								}
-								<button type="button" onClick={callSignInForm}>Sign In</button>
-								<button type="button" onClick={callRegisterForm}>Register</button>
-							</Form>
-
-						)
-					}
+							</RequestSubTypeField>
+							<div className="input-feedback">
+								<ErrorMsg
+									errormessage={rest.formik.errors.subRequestType}
+									touched={rest.formik.touched.subRequestType} />
+							</div>
+						</div>
+						: null
 				}
-			</Formik>
+				{
+
+					/* water and Sewer Issues --- Sewer Issues */
+					/* water and Sewer Issues --- Sewer Issues */
+					/* water and Sewer Issues --- Sewer Issues*/
+
+					(rest.formik.values['requestType'].toLowerCase() === requestType_WaterandSewerIssues.toLowerCase())
+						&& (rest.formik.values['subRequestType'].toLowerCase() === subCategory_SewerIssues.toLowerCase() ||
+							rest.formik.values['subRequestType'].toLowerCase() === subCategory_StormWaterIssues.toLowerCase() ||
+							rest.formik.values['subRequestType'].toLowerCase() === subCategory_WaterSupplyIssues.toLowerCase()
+						) ? notes : null
+				}
+
+
+				{
+					/* Can or lid lost or damaged*/
+					/* Property damage during collection */
+					/* Recycling not collected */
+					/* Request to start new collection */
+					/* Trash not collected */
+					/* Yard waste not collected*/
+
+					(rest.formik.values['requestType'].toLowerCase() === requestType_TrashRecycleIssue.toLowerCase())
+						&& (rest.formik.values['subRequestType'].toLowerCase() === subCategory_CanOrLidLostDamaged.toLowerCase() ||
+							rest.formik.values['subRequestType'].toLowerCase() === subCategory_PropertyDamangeDuringCollecttion.toLowerCase() ||
+							rest.formik.values['subRequestType'].toLowerCase() === subCategory_RecyclingNotCollected.toLowerCase() ||
+							rest.formik.values['subRequestType'].toLowerCase() === subCategory_RequestToStartNewCollection.toLowerCase() ||
+							rest.formik.values['subRequestType'].toLowerCase() === subCategory_TrashNotCollected.toLowerCase() ||
+							rest.formik.values['subRequestType'].toLowerCase() === subCategory_YardWasteNotCollected.toLowerCase()
+						) ? notes : null
+
+				}
+
+
+
+
+				{ /* Roads and Sidewalks Issue -- icy condition*/
+					(rest.formik.values['requestType'].toLowerCase() === requestType_RoadSidewalkIssue.toLowerCase()
+						&& rest.formik.values['subRequestType'].toLowerCase() === subCategory_IcyConditions.toLowerCase()) ? notes : null
+
+				}
+				{/* Pets and Animal Issue - Other animal complaint */
+
+					(rest.formik.values['requestType'].toLowerCase() === requestType_petAndAnimalIssue.toLowerCase()
+						&& rest.formik.values['subRequestType'].toLowerCase() === petAndAnimalIssueID_OtherAnimalComplaint.toLowerCase()) ? notes
+						: null
+				}
+				{/* Website Issue - Other website problem */
+
+					(rest.formik.values['requestType'].toLowerCase() === requestType_WebSiteIssue.toLowerCase()
+						&& rest.formik.values['subRequestType'].toLowerCase() === subCategory_OtherWebsiteProblem.toLowerCase()) ? notes
+						: null
+				}
+				{
+					rest.formik.values['requestType'] === requestType_petAndAnimalIssue && rest.formik.values['subRequestType'] !== '' ?
+						<div>
+							<label htmlFor="petType"
+								className={
+									rest.formik.errors.petType && rest.formik.touched.petType ? "input-feedback" : "text-label"}
+							>Pet Type</label>
+							<RequestPetTypeField
+								component="select"
+								name="petType"
+								formikProps={rest}
+								onChange={handleServicePetChange}
+								//value={props.formik.values.name}
+								className={rest.formik.errors.petType && rest.formik.touched.petType ? "text-select error" : null}
+							>
+								<option key='default' value=''>--Please select a pet type--</option>
+								{PetTypes.map(petType => (
+									<option key={petType.id} value={petType.name}>{petType.name}</option>
+								))}
+							</RequestPetTypeField>
+							<div className="input-feedback">
+								{<ErrorMsg
+									errormessage={rest.formik.errors.petType}
+									touched={rest.formik.touched.petType} />}
+							</div>
+						</div>
+						: null
+				}
+				{
+
+					(rest.formik.values['subRequestType'] !== '' && rest.formik.values['petType'].toLowerCase() === petType_Others) ?
+						<div>
+							<label htmlFor="otherAnimalTypes"
+								className={
+									rest.formik.errors.otherAnimalTypes && rest.formik.touched.otherAnimalTypes ? "input-feedback" : "text-label"}
+							>Other pet type
+							</label>
+							<GenericTypeField
+								component="select"
+								name="otherAnimalTypes"
+								formikProps={rest}
+								onChange={handleOtherPetTypeChange}
+								//	value={props.formik.values.name}
+								className={rest.formik.errors.otherAnimalTypes && rest.formik.touched.otherAnimalTypes ? "text-select error" : null}
+							>
+								<option key='default' value=''>--Please select an "other" pet type--</option>
+								{OtherAnimalTypes.map(OtherAnimalType => (
+									<option key={OtherAnimalType.id} value={OtherAnimalType.name}>{OtherAnimalType.name}</option>
+								))}
+
+							</GenericTypeField>
+
+							<div className="input-feedback">
+								{<ErrorMsg
+									errormessage={rest.formik.errors.otherAnimalTypes}
+									touched={rest.formik.touched.otherAnimalTypes} />}
+							</div>
+						</div>
+						: null
+				}
+				{
+					((rest.formik.values['requestType'] === requestType_petAndAnimalIssue)
+						&& rest.formik.values['subRequestType'] !== '')
+						&& checkPetType(rest.formik.values['petType']) ?
+						<div>
+							<label htmlFor="sexType"
+								className={
+									rest.formik.errors.sexType && rest.formik.touched.sexType ? "input-feedback" : "text-label"}
+							>Pet Sex (optional)
+							</label>
+							<GenericTypeField
+								component="select"
+								name="sexType"
+								formikProps={rest}
+								onChange={handlePetSexChange}
+								//value={props.formik.values.name}
+								className={rest.formik.errors.sexType && rest.formik.touched.sexType ? "text-select error" : null}
+							>
+								<option key='default' value=''>--Please select a pet sex--</option>
+								{animalSex.map(petSex => (
+									<option key={petSex.id} value={petSex.name}>{petSex.name}</option>
+								))}
+							</GenericTypeField>
+
+							<div className="input-feedback">
+								{<ErrorMsg
+									errormessage={rest.formik.errors.sexType}
+									touched={rest.formik.touched.sexType} />}
+							</div>
+						</div>
+						: null
+				}
+
+				{
+
+					(rest.formik.values['requestType'] === requestType_petAndAnimalIssue
+						&& rest.formik.values['subRequestType'] !== '')
+						&& (rest.formik.values['petType'].toLowerCase() === petTypeCat || rest.formik.values['petType'].toLowerCase() === petTypeDog) ?
+						<div>
+							<label htmlFor="animalColorType"
+								className={
+									rest.formik.errors.animalColorType && rest.formik.touched.animalColorType ? "input-feedback" : "text-label"}
+							>Primary Animal Color
+							</label>
+
+							<GenericTypeField
+								component="select"
+								name="animalColorType"
+								formikProps={rest}
+								onChange={handleAnimalColorChange}
+								//value={props.formik.values.name}
+								className={rest.formik.errors.animalColorType && rest.formik.touched.animalColorType ? "text-select error" : null}
+							>
+								<option key='default' value=''>--Please select the primary color of the animal--</option>
+
+								{AnimalColors.map(animalColorType => (
+									<option key={animalColorType.id} value={animalColorType.name}>{animalColorType.name}</option>
+								))}
+
+							</GenericTypeField>
+							<div className="input-feedback">
+								{<ErrorMsg
+									errormessage={rest.formik.errors.animalColorType}
+									touched={rest.formik.touched.animalColorType} />}
+							</div>
+						</div>
+						: null
+
+				}
+
+				<br />
+				{
+					((rest.formik.values['requestType'] === requestType_petAndAnimalIssue && rest.formik.values['subRequestType'] !== '')
+						&& (rest.formik.values['petType'].toLowerCase() === petTypeCat || rest.formik.values['petType'].toLowerCase() === petTypeDog)) ?
+						<div>
+							<label htmlFor="animalBreed"
+								className={
+									rest.formik.errors.animalBreed && rest.formik.touched.animalBreed ? "input-feedback" : "text-label"}
+							>Primary Animal Breed(optional)
+							</label>
+							<Field
+								component="select"
+								name="animalBreedType"
+								onChange={handleAnimalBreedChange}
+								value={props.formik.values.name}
+								className={rest.formik.errors.animalBreedType && rest.formik.touched.animalBreedType ? "text-select error" : null}
+							>
+								<option key='default' value=''>--Please select the primary breed of the animal--</option>
+								{animalSubCategories.map(animalBreedType => (
+									<option key={animalBreedType.id} value={animalBreedType.name}>{animalBreedType.name}</option>
+								))}
+							</Field>
+							<div className="input-feedback">
+								{<ErrorMsg
+									errormessage={rest.formik.errors.animalBreedType}
+									touched={rest.formik.touched.animalBreedType} />}
+
+							</div>
+						</div>
+						: null
+
+				}
+				<Field type="hidden" name="requestTypeID" />
+				<Field type="hidden" name="subRequestTypeID" />
+				<Field type="hidden" name="petTypeID" />
+				<Field type="hidden" name="animalBreedID" />
+				<Field type="hidden" name="sexTypeID" />
+				<Field type="hidden" name="animalColorTypeID" />
+				<Field type="hidden" name="otherAnimalTypesID" />
+				<Field type="hidden" name="descriptionID" />
+				<Field type="hidden" name="streeAddressID" />
+				<Field type="hidden" name="cityID" />
+				<Field type="hidden" name="zipCodeID" />
+
+				{(rest.formik.values['ContactID'] === '') ?
+					<div>
+						<button type="button" onClick={callSignInForm} disabled={displayButton}>
+							Sign In
+						</button>
+						<button type="button" onClick={callRegisterForm} disabled={displayButton}>Register</button>
+						<Model />
+
+					</div> : <button type="button" disabled={displayButton} onClick={goToNextPage}>Next</button>
+
+
+				}
+
+			</Form>
+
+
 		</FormContainer>
 
 	);
@@ -377,3 +610,21 @@ const ServiceRequestForm = props => {
 
 export default connect(ServiceRequestForm);
 
+
+
+
+
+/* 	(rest.formik.values['requestType'].toLowerCase() === requestType_WaterandSewerIssues.toLowerCase()
+		&& rest.formik.values['subRequestType'].toLowerCase() === subCategory_SewerIssues.toLowerCase()) ?
+		<div>
+			<p>
+				Issues such as missing manhole covers, sewer backups, overflows or odors require immediate attention, and therefore cannot be reported online.
+		
+			</p>
+			<p>
+				Please call the Department of Public Works at 410-887-7415 to ensure we obtain the necessary information to address the issue as soon as possible.
+			</p>
+
+		</div>
+		: null
+ */
