--- src/components/SubmitButton.js
@@ -1,5 +1,0 @@
-const SubmitButton = () => null;
-
-SubmitButton.displayName = 'SubmitButton';
-
-export default SubmitButton;
